protect delete admin by changing value

+++ admin_users.php
@@ -8,8 +8,10 @@ if ($_SESSION['admin'] !== true) {
 }
 if ($_SERVER['REQUEST_METHOD'] === "POST") {
 	if ($_POST['type'] === "delete") {
-		delete_user("./database/users", $_POST['username']);
-		delete_user("./database/passwd", $_POST['username']);
+	    if (!is_admin_user($_POST['username'])) {
+            delete_user("./database/users", $_POST['username']);
+            delete_user("./database/passwd", $_POST['username']);
+        }
     }
     header('Location: admin_users.php');
 }
